Update Tweak.xm
I got your PM about the current app receiving keystrokes instead of SpringBoard. I think that the problem may be that the SBAppWindow (Second window of SpringBoard) is not the key window, instead the current application's window it. 

I haven't tested this yet, this is just a suggestion. If this does end up working, make sure that you also make the SBAppWindow resign being the key window when Searchlight is closed. Hope that this helps!